+++ src/main.js
@@ -1,13 +1,15 @@
 import Vue from "vue";
-import App from "./App.vue";
-import router from "./router";
 import store from "./store";
+import App from "./App.vue";
+import { Router, i18n } from "./plugins";
+import "./assets/scss/styles.scss";
 
 Vue.config.productionTip = false;
 
 new Vue({
-    router,
+    Router,
     store,
+    i18n,
     render(h) {
         return h(App);
     },