+++ src/plugins/bootstrap/index.js
@@ -1,0 +1,4 @@
+import "jquery";
+import "popper.js";
+import "bootstrap";
+import "bootstrap/dist/css/bootstrap.min.css";