+++ src/plugins/index.js
@@ -1,0 +1,7 @@
+import "./APIs";
+import "./bootstrap";
+import "./error";
+import Router from "./router";
+import i18n from "./vue-i18n";
+
+export { Router, i18n };