+++ src/plugins/router/index.js
@@ -1,0 +1,19 @@
+import Vue from "vue";
+import VueRouter from "vue-router";
+import Home from "../../components/Home";
+
+Vue.use(VueRouter);
+
+const routes = [{
+    path: "/",
+    name: "Home",
+    component: Home,
+}];
+
+const Router = new VueRouter({
+    mode: "history",
+    base: process.env.BASE_URL,
+    routes,
+});
+
+export default Router;