+++ CHANGELOG.md
@@ -1,5 +1,26 @@
 # 📈 LIST OF CHANGES FOR WEREWOLVES ASSISTANT WEB
 
+## 0.1.0 (2020-06-??)
+
+### 🚀 New features
+
+* Home page with play, log in and stats buttons.
+
+### 📦 Packages
+
+* `@fortawesome/fontawesome-free` installed with version `5.13.1`.
+* `axios` installed with version `0.19.2`.
+* `bootstrap` installed with version `4.5.0`.
+* `jquery` installed with version `3.5.1`.
+* `jsonwebtoken` installed with version `8.5.1`.
+* `node-sass` installed with version `4.14.1`.
+* `popper.js` installed with version `1.16.1`.
+* `sass-loader` installed with version `8.0.2`.
+* `vue-i18n` installed with version `8.18.2`.
+* `vue-toasted` installed with version `1.1.28`.
+
+---
+
 ## 0.0.0 (2020-06-20)
 
 ### 🚀 New features